requirements.txt
PyGithub==1.57
requests~=2.28.1